check htmlcompressor-1.5.3.jar exist.

+++ install.sh
@@ -81,7 +81,7 @@ which sqwish 1> /dev/null 2>&1 || npm install -g sqwish
 
 # fetch htmlcompressor java library
 [ -d ~/.htmlcompressor ] || mkdir ~/.htmlcompressor
-wget https://github.com/appleboy/html5-template-engine/raw/master/build/htmlcompressor-1.5.3.jar -O ~/.htmlcompressor/htmlcompressor-1.5.3.jar
+[ -f ~/.htmlcompressor/htmlcompressor-1.5.3.jar ] || wget https://github.com/appleboy/html5-template-engine/raw/master/build/htmlcompressor-1.5.3.jar -O ~/.htmlcompressor/htmlcompressor-1.5.3.jar
 
 # install ruby image optimize tool
 install_image_optim
